### 2020-05-25 #2252 @whedon
[PRE REVIEW]: CancerSim: A python3 package for stochastic cancer simulation in 2D
**Submitting author:** @CFGrote (<a href="http://orcid.org/0000-0002-2579-5546">Carsten Fortmann-Grote</a>)
**Repository:** <a href="https://github.com/mpievolbio-scicomp/cancer_sim" target ="_blank">https://github.com/mpievolbio-scicomp/cancer_sim</a>
**Version:** v1.0.0
