### 2020-06-15 #2252 @CFGrote
MIT License added to master
