### 2020-06-16 #2252 @pdebuyl
@jenzopr , @hannahbrucemacdonald , @blakeaw could you review the manuscript "CancerSim: A python3 package for stochastic cancer simulation in 2D" for the Journal of Open Source Software? (two reviewers needed).
