### 2020-06-24 #2252 @pdebuyl
@CFGrote while we are waiting for the confirmation of a second reviewer, I have a comment already. The paper includes a full example for the parameter file. JOSS articles are intended to remain concise and this content is best left in the dedicated documentation.
