### 2020-06-29 #2252 @pdebuyl
@dirmeier would you review the article "CancerSim: A python3 package for stochastic cancer simulation in 2D" for the Journal of Open Source Software?
