### 2020-07-04 #2436 @whedon
[REVIEW]: CancerSim: A python3 package for stochastic cancer simulation in 2D
**Submitting author:** @CFGrote (<a href="http://orcid.org/0000-0002-2579-5546">Carsten Fortmann-Grote</a>)
**Repository:** <a href="https://github.com/mpievolbio-scicomp/cancer_sim" target ="_blank">https://github.com/mpievolbio-scicomp/cancer_sim</a>
**Version:** 2.0.1
