### 2020-07-14 #2436 @blakeaw
Hi @CFGrote, based on your software's commit history it looks like you are the only author that made direct contributions to the code itself. Would you mind briefly stating what contributions the other authors listed in the paper made to the project?
