### 2020-07-14 mpievolbio-scicomp/cancer_sim#5 @blakeaw
Missing pip install of casim package in Installation directions?
I followed the steps to install casim using a conda env, but upon running the tests I got the following error:

Traceback (most recent call last):
  File "casim_test.py", line 4, in <module>
    from casim import casim
ModuleNotFoundError: No module named 'casim'

I think the Installation instructions may be missing an explicit step. I tried running `pip install .` from the cancer_sim directory and casim installed in my environment and then the tests ran OK.
