### 2020-07-15 mpievolbio-scicomp/cancer_sim#5 @jenzopr
You might also want to create a console script entrypoint in `setup.py` and then uploade to pip and/or create a conda package. Installation instructions and access to the high level function would simplify drastically.
