### 2020-07-15 mpievolbio-scicomp/cancer_sim#6 @jenzopr
Example fails bc mut_prob is not found in params.py
Hi! I tried to follow the example provided in the README and as well in the docs at readthedocs.

However, the examples (`python -m casim.casim 1`) fail with

```
Traceback (most recent call last):
  File "/home/jpreuss/.conda/envs/casim/lib/python3.7/runpy.py", line 193, in _run_module_as_main
    "__main__", mod_spec)
  File "/home/jpreuss/.conda/envs/casim/lib/python3.7/runpy.py", line 85, in _run_code
    exec(code, run_globals)
  File "/home/jpreuss/cancer_sim/casim/casim.py", line 1112, in <module>
    sys.exit(main(arguments))
  File "/home/jpreuss/cancer_sim/casim/casim.py", line 1021, in main
    mutation_probability = params.mut_prob,
AttributeError: module 'params' has no attribute 'mut_prob'
```

I had to change `mutation_probabilty` in line 20 of [params.py](https://github.com/mpievolbio-scicomp/cancer_sim/blob/master/casim/params.py) to `mut_prob`.

Both, the REAMDE and the docs at readthedocs contain the path to `test/params.py` as params file. The source distribution does not contain the file at this location (rather at `casim/params.py`). Maybe update the docs and as an additional suggestion, I could imagine that a `params` parameter to `casim.casim` would help to point the software to the parameter file that should be used (instead of relying on the presence of a file called `params.py`).

Best,
Jens
