### 2020-07-15 #2436 @jenzopr
Problems with reproducing the Example usage from README and the docs: Issue https://github.com/mpievolbio-scicomp/cancer_sim/issues/6
