### 2020-07-15 mpievolbio-scicomp/cancer_sim#7 @jenzopr
Community guidelines missing
Hi!

This issue is part of the JOSS review at https://github.com/openjournals/joss-reviews/issues/2436

The review requires **Community guidelines**: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Those are missing in the source distribution or docs at readthedocs.
