### 2020-07-15 mpievolbio-scicomp/cancer_sim#5 @blakeaw
This issue is part of the JOSS review at https://github.com/openjournals/joss-reviews/issues/2436
