### 2020-07-15 mpievolbio-scicomp/cancer_sim#6 @blakeaw
I got the same error when I tried to run the examples from within the cancer_sim directory.

However, when I switched to a different directory and ran them they seemed to work ok. Maybe this a problem too? I'm not sure at the moment where the main function is reading the params from in this case.
