### 2020-07-15 mpievolbio-scicomp/cancer_sim#6 @blakeaw
As suggested by @jenzopr, I think having a required `params` flag for `casim.casim` is probably the better way to point it to the params file/module. With this approach, you could also give users a little more flexibility in naming their parameter modules, as opposed to just having them named `params.py`.
