### 2020-07-15 mpievolbio-scicomp/cancer_sim#6 @blakeaw
I'll also note that from the current documentation it wasn't entirely clear to me at first that casim.casim by default imports parameters from the param.py, and that it tries to find it in the current working directory. I think you may need to make this more explicit in the documentation. However, this comment would probably be moot if you implement the `params` flag/parmameter for `casim.casim` calls.
