### 2020-07-15 mpievolbio-scicomp/cancer_sim#8 @blakeaw
Additional high level documentation for simulation outputs
Hello,

This issue is part of the JOSS review at https://github.com/openjournals/joss-reviews/issues/2436

I'd recommend adding some additional high-level documentation on what the simulation outputs are for runs with `casim.casim`. Currently, it looks like the only related documentation is the following under the examples: 'Results will be stored in the newly created directory `sim_out/`.'  However, there doesn't appear to be any documentation on what the expected output file-structure is in the `sim_out` out directory or what information each output file contains. It may also be helpful for new users if you give some hints on how to follow up with analyzing those outputs - assuming it's not entirely self-explanatory in the output file descriptions.
