### 2020-07-15 mpievolbio-scicomp/cancer_sim#9 @blakeaw
Low level API auto generated documentation missing from docs/readthedocs
Hello,

This issue is part of the JOSS review at https://github.com/openjournals/joss-reviews/issues/2436

Docs: https://cancer-sim.readthedocs.io/en/latest/index.html

It looks like there may have been a problem in building your low-level API documentation. I see in the sphinx conf.py that you imported the `autodoc` extension, but I can't find any module or class-level docs autogenerated from docstrings. The Reference Manual page (https://cancer-sim.readthedocs.io/en/latest/refman.html) is also blank on readthedocs, which is where I thought the low-level auto generated docs would probably be located.
