### 2020-07-15 #2436 @blakeaw
Recommendation for additional documentation of simulation outputs: https://github.com/mpievolbio-scicomp/cancer_sim/issues/8
