### 2020-07-15 mpievolbio-scicomp/cancer_sim#8 @blakeaw
Similarly, for outputs generated by `casim.CancerSimulator.run`.
