### 2020-07-17 mpievolbio-scicomp/cancer_sim#8 @blakeaw
Ok, I re-read the Background in the README and I see that towards the end it says "After the simulation, the tumour matrix, and the lists of lineages and frequencies of each mutation in the tumour are exported to files." So you can probably just restate this under the high-level functionality description along with a more detailed layout of the output directory structure and filenames generated when you run casim. And you might also add this information to the quickstart example too, just to remind users again what to expect as outputs in case they missed it elsewhere (like I did).
