### 2020-07-17 mpievolbio-scicomp/cancer_sim#10 @blakeaw
Documentation doesn't say what the example and Quickstart cases are simulating
Hello,

This issue is part of the JOSS review at https://github.com/openjournals/joss-reviews/issues/2436

For the example you provided in the README and also for the one in the Quickstart it looks like there isn't information about what cases they are actually simulating. For example, are the parameters modeling a specific tumor type? Or is this essentially just a toy-model set of parameters? I think it might be good to mention this in the documentation/examples.
