### 2020-07-21 mpievolbio-scicomp/cancer_sim#11 @jenzopr
Software paper: Review
This issue is part of the JOSS review at openjournals/joss-reviews#2436

## Software paper review

The work of Opasic and colleagues introduces CancerSim, a python package to study the evolution of somatic mutations in cancer cells, by introducing a simple simulation procedure on a 2D lattice and along discrete time steps. The package features key parameters, like cell division and mutation probabilities, tracks parent-daugther cell relationships and allows for adjustment of cell fitness of mutated cells. It therefore resembles an abstract model of superficially spreading tumors.

Although the paper consists of a well-written summary and an in-depth algorithm outline, description of key parameters remains superficial and applicability of the package in theoretical modelling is only vaguely delineated.

### Major comments

- [x] The second paragraph of the **Summary** dives straight into specialist concepts like "cancer evolution", "somatic mutations", "genetic heterogeneity" and "sequencing". To enhance understanding for a diverse, non-specialist audience, please enhance the short conceptual introduction in the first paragraph, picking up those concepts. (Software paper: Summary)
- [x] There exists some discrepancy between parameter description and parameter naming (e.g. the number of divisions per generation vs. `div_probability`, number of division for cells with mutation vs. `fittnes_advantage_div_prob`). Please make sure to revise parameter names to indicate e.g. if a probability or absolute number is required, and include descriptions of which and how parameters interplay or depend on each other. (Documentation: Functional documentation)
- [x] A description of simulation output is missing in the paper and the online help. Having a detailed description of the output files lets readers quickly judge wheter or not a software will be suitable for their needs. (Documentation: Functional documentation)
- [x] Having a detailed description of simulation output will also help readers to envision how the software can be applied in theoretical modelling and, importantly, which related research questions can be addressed e.g. in follow-up studies. Please delineate the softwares potential in much more detail, by suggestions or elaboration on possible downstream analysis. (Software paper: Statement of need)

- [x] The first paragraph mentions cancer cell dormancy as an example for beneficial use of theoretical models to understand carcinogenesis. Unfortunately, CancerSim does not allow cells to be simulated dormant, which renders the example detrimental to the papers statement.
- [x] State of the field: Please include a short statement of which other tools are available for somatic mutation simulation and briefly compare features/drawbacks.
- [x] The summary states that simulated tumors can be subjected for multi-region sampling. Along my last major comment, please elaborate how this can be done in light of algorithm output.
- [x] In the fifth paragraph of the summary, please change "In each simulation step, every tumour cell in the tumour that has an unoccupied neighbour" to "In each simulation step, every tumour cell **on the lattice** that has an unoccupied neighbour"
- [x] The fifth paragraph of the summary also states that variability in fitness can be introduced for some cells. Please be more specific: How and for which cells can varaibility in fitness be introduced? How can beneficial and deleterious mutation rates be encoded?
- [x] Can other sources of sequencing noise be introduced (e.g. sequencing errors at non-mutated sites)?
