### 2020-07-21 #2436 @jenzopr
I reviewed the paper in https://github.com/mpievolbio-scicomp/cancer_sim/issues/11 and will grade the submission in category 3) Major revision.
