### 2020-07-27 46e9fc6
Added community guidelines in response to gh issue #7.
