### 2020-07-27 9904e76
Adding cwd to environment. Fixes gh issue #5.
