### 2020-07-27 d77927b
Adding cwd to requirements. Fixes gh issue #5.
