+++ paper.md
@@ -142,34 +142,34 @@ programmatically via the `CancerSimulationParameters` class. The file
 matrix_size                      = 100
 
 # Number of generations to simulate.
-num_of_generations              = 20
+number_of_generations              = 20
 
 # Number of divisions per generation.
-div_probability                 = 1
+division_probability                 = 1
 
 # Number of division for cells with mutation.
-fittnes_advantage_div_prob      = 1
+adv_mutant_division_probability      = 1
 
 # Fraction of cells that die per generation.
-dying_fraction                   = 0.1
+death_probability                   = 0.1
 
 # Fraction of cells with mutation that die per generation.
-fitness_advantage_death_prob    = 0.0
+adv_mutant_death_probability    = 0.0
 
 # Rate of mutations.
-mutation_probabilty              = 1
+mutation_probability              = 1
 
 # Mutation probability for the adv. cells.
-advantageous_mut_prob           = 1
+adv_mutant_mutation_probability           = 1
 
 # Number of mutations per cell division.
-mut_per_division                = 10
+number_of_mutations_per_division                = 10
 
 # Time after which adv. mutations occur.
-time_of_adv_mut                 = 10
+adv_mutation_interval                 = 10
 
 # Number of mutations present in first cancer cell.
-num_of_clonal                   = 150
+number_of_initital_mutations                   = 150
 
 # Tumour multiplicity.
 tumour_multiplicity             = None
