+++ test/casim_test.py
@@ -50,59 +50,59 @@ class CancerSimulatorParametersTest(unittest.TestCase):
 
         parameters = CancerSimulatorParameters()
 
-        self.assertEqual(parameters.matrix_size,                          10  )
-        self.assertEqual(parameters.number_of_generations,                 2  )
-        self.assertEqual(parameters.division_probability,                  1  )
-        self.assertEqual(parameters.advantageous_division_probability,     1  )
-        self.assertEqual(parameters.death_probability,                     0  )
-        self.assertEqual(parameters.advantageous_death_probability,   0.0)
-        self.assertEqual(parameters.mutation_probability,                         0.8)
-        self.assertEqual(parameters.advantageous_mutation_probability,     1  )
-        self.assertEqual(parameters.mutations_per_division,                1  )
-        self.assertEqual(parameters.time_of_advantageous_mutation,     50000  )
-        self.assertEqual(parameters.number_of_clonal,                     1   )
-        self.assertEqual(parameters.tumour_multiplicity,              'single')
-        self.assertEqual(parameters.read_depth,                         100  )
-        self.assertEqual(parameters.sampling_fraction,                         0.0  )
+        self.assertEqual(parameters.matrix_size, 10)
+        self.assertEqual(parameters.number_of_generations, 2)
+        self.assertEqual(parameters.division_probability, 1)
+        self.assertEqual(parameters.adv_mutant_division_probability, 1)
+        self.assertEqual(parameters.death_probability, 0)
+        self.assertEqual(parameters.adv_mutant_death_probability, 0.0)
+        self.assertEqual(parameters.mutation_probability, 0.8)
+        self.assertEqual(parameters.adv_mutant_mutation_probability, 1)
+        self.assertEqual(parameters.mutations_per_division, 1)
+        self.assertEqual(parameters.adv_mutation_interval, 50000)
+        self.assertEqual(parameters.number_of_initital_mutations, 1)
+        self.assertEqual(parameters.tumour_multiplicity, 'single')
+        self.assertEqual(parameters.read_depth, 100)
+        self.assertEqual(parameters.sampling_fraction, 0.0)
         self.assertTrue(parameters.plot_tumour_growth)
         self.assertTrue(parameters.export_tumour)
 
     def test_shaped_constructor (self):
         """ Test initialization with arguments. """
 
-        parameters = CancerSimulatorParameters(
-                                matrix_size =                           20  ,
-                                number_of_generations =                 10  ,
-                                division_probability =                   0.5,
-                                advantageous_division_probability =      0.3,
-                                death_probability =                      0.1,
-                                advantageous_death_probability =    0.4,
-                                mutation_probability =                          0.2,
-                                advantageous_mutation_probability =      0.8,
-                                mutations_per_division =                10  ,
-                                time_of_advantageous_mutation =      30000  ,
-                                number_of_clonal =                       2  ,
-                                tumour_multiplicity =                'single',
-                                read_depth = 200,
-                                sampling_fraction = 0.3,
-                                export_tumour = False,
-                                plot_tumour_growth = False,
+        parameters=CancerSimulatorParameters(
+                                matrix_size=20  ,
+                                number_of_generations=10  ,
+                                division_probability=0.5,
+                                adv_mutant_division_probability=0.3,
+                                death_probability=0.1,
+                                adv_mutant_death_probability=0.4,
+                                mutation_probability=0.2,
+                                adv_mutant_mutation_probability=0.8,
+                                mutations_per_division=10  ,
+                                adv_mutation_interval=30000  ,
+                                number_of_initital_mutations=2  ,
+                                tumour_multiplicity='single',
+                                read_depth=200,
+                                sampling_fraction=0.3,
+                                export_tumour=False,
[…]
