+++ casim/casim.py
@@ -41,22 +41,22 @@ class CancerSimulatorParameters(object):
     """
 
     def __init__(self,
-                 matrix_size                         = None,
-                 number_of_generations               = None,
-                 division_probability                = None,
-                 advantageous_division_probability   = None,
-                 death_probability                   = None,
-                 advantageous_death_probability      = None,
-                 mutation_probability                       = None,
-                 advantageous_mutation_probability   = None,
-                 mutations_per_division              = None,
-                 time_of_advantageous_mutation       = None,
-                 number_of_clonal                    = None,
-                 tumour_multiplicity                 = None,
-                 read_depth                          = None,
-                 sampling_fraction                   = None,
-                 plot_tumour_growth                  = None,
-                 export_tumour                       = None,
+                 matrix_size=None,
+                 number_of_generations=None,
+                 division_probability=None,
+                 adv_mutant_division_probability=None,
+                 death_probability=None,
+                 adv_mutant_death_probability=None,
+                 mutation_probability=None,
+                 adv_mutant_mutation_probability=None,
+                 mutations_per_division=None,
+                 adv_mutation_interval=None,
+                 number_of_initital_mutations=None,
+                 tumour_multiplicity=None,
+                 read_depth=None,
+                 sampling_fraction=None,
+                 plot_tumour_growth=None,
+                 export_tumour=None,
                 ):
         """
         Construct a new CancerSimulationParameters object.
@@ -70,29 +70,29 @@ class CancerSimulatorParameters(object):
         :param division_probability: The probability for a cell division to occur during one generation.
         :type  division_probability: float (0.0 <= division_probability <= 1.0)
 
-        :param advantageous_division_probability: The probability for the division of a cell with advantageous mutation to occur during one generation.
-        :type  advantageous_division_probability: float (0.0 <= division_probability <= 1.0)
+        :param adv_mutant_division_probability: The probability for the division of a cell with advantageous mutation to occur during one generation.
+        :type  adv_mutant_division_probability: float (0.0 <= division_probability <= 1.0)
 
         :param death_probability: The probability for a cell to die during one generation.
         :type  death_probability: float (0.0 <= division_probability <= 1.0)
 
-        :param advantageous_death_probability: The probability for a cell with advantageous mutation to die during one generation.
-        :type  advantageous_death_probability: float (0.0 <= division_probability <= 1.0)
+        :param adv_mutant_death_probability: The probability for a cell with advantageous mutation to die during one generation.
+        :type  adv_mutant_death_probability: float (0.0 <= division_probability <= 1.0)
 
         :param mutation_probability: The probalitiy of mutation.
         :type  mutation_probability: float (0.0 <= division_probability <= 1.0)
 
-        :param advantageous_mutation_probability: The rate for an advantageous mutation to occur during one generation.
-        :type  advantageous_mutation_probability: float (0.0 <= division_probability <= 1.0)
+        :param adv_mutant_mutation_probability: The rate for an advantageous mutation to occur during one generation.
+        :type  adv_mutant_mutation_probability: float (0.0 <= division_probability <= 1.0)
 
         :param mutations_per_division: The number of mutations per division
         :type  mutations_per_division: int
 
-        :param time_of_advantageous_mutation: The number of generations after which an advantageous mutation can occur.
[…]

[…]
