### 2020-07-27 mpievolbio-scicomp/cancer_sim#6 @CFGrote
i fixed the original issue in PR #14 but i'll also refactor the handling of `params.py` taking into account the suggestions above.
