### 2020-07-27 54b8868
Fixing tests after refactoring the CLI as suggested by reviewers in #6.
