+++ README.md
@@ -77,7 +77,7 @@ After the simulation, the tumour matrix, and the lists of lineages and
 frequencies of each mutation in the tumour are exported to files.
 Furthermore, the virtual tumour can be sampled and a histogram over the
 frequency of mutations will be visualised. Alternatively, a saved tumour
-can be loaded from file and then subjected to the sampling process.
+can be loaded from file and then be subjected to the sampling process.
 
 Installation
 ------------
@@ -161,9 +161,10 @@ The test suite is automatically run after each commit to the code base.
 Results are published on
 [travis-ci.org](https://travis-ci.org/mpievolbio-scicomp/cancer_sim).
 
-High--level functionality
+High-level functionality
 -------------------------
 
+### Setting the simulation parameters
 The parameters of the cancer simulation are given via a python module or
 programmatically via the `CancerSimulationParameters` class. A
 documented example `params.py` is included in the source code (under
@@ -218,47 +219,77 @@ documented example `params.py` is included in the source code (under
     # Export the tumour growth data to file.
     export_tumour = True
 
+Here, we simulate a 2D tumour on a 100x100 grid for a total of 20 generations.
+On average, both healthy and mutant cells divide once per generation. The first
+cancer cell carries 150 mutations.
+Mutant cells with advantageous mutations live on forever while healthy cells die with a rate of 0.1 per generation.
+Both healthy and mutant cells aquire 10 new mutations in
+each generation with a certainty of 100%. Every 10th generation, a mutation is
+advantageous (i.e. cancerogenic).
+
+90% of all cells are sampled on a random basis and sequenced with a read depth
+of 100. The data is written to disk and a plot showing the tumour growth in
+generated.
+ 
 The simulation is started from the command line. The syntax is
 
-    $> python -m casim.casim [-h] [-o DIR] seed
+    $> python -m casim.casim [-h] [-s SEED] [-p PARAMS] [-o DIR]
 
-The mandatory command line argument `seed` is the random seed. Using the
+`SEED` is the random seed. Using the
 same seed in two simulation runs with identical parameters results in
-identical results. This feature may be used for testing and debugging.
-The optional argument `DIR` specifies the directory where to store the
+identical results. If not given, `SEED` defaults to 1. `PARAMS` should point to
+a python parameter file. If not given, it defaults to `params.py` in the current
+working directory. If that file does not exist, default parameters are assumed.
+`DIR` specifies the directory where to store the
 simulation log and output data. If not given, output will be stored in
 the directory `casim_out` in the current directory. For each seed, a
 subdirectory `cancer_SEED` will be created. If that subdirectory already
 exists because an earlier run used the same seed, the run will abort.
 This is a safety catch to avoid overwriting data from previous runs.
 
-### Example 1
+### Run the example
 
-    $> python -m casim.casim 1
+    $> python -m casim.casim 1 -p casim/params.py -o cancer_sim_example
 
-### Example 2
+### Output
+After the run has finished, you should find the results in
+`cancer_sim_example/cancer_1/simOutput`.
 
-    $> mkdir sim_out
-    $> python -m casim.casim -o sim_out 2
+    $> ls cancer_sim_example/cancer_1/simOutput
+    death_list.p     mtx.p        mut_container.p              sample_out_502_488.txt
+    growthCurve.pdf  mtx_VAF.txt  sampleHistogram_502_488.pdf  wholeTumourVAFHistogram.pdf
+   
+Files with the extension `.p` are binary files (python pickles) needed to
+restart a simulation. Let's take a look at the `.txt`  files. They contain the simulation output:
[…]
