### 2020-07-28 mpievolbio-scicomp/cancer_sim#10 @CFGrote
@lopasic, can you please review PR #15, in particular check scientific
background of examples both in README and in the notebook.

thanks! c
