### 2020-07-28 mpievolbio-scicomp/cancer_sim#9 @CFGrote
hi,
i cannot reproduce your observation.

after pushing a few times to master, the doc correctly builds at the
mentioned link https://cancer-sim.readthedocs.io/en/latest/index.html
and the reference manual is where it should be at
https://cancer-sim.readthedocs.io/en/latest/refman.html.

regards,
carsten
