### 2020-07-28 61220b7
Small amendment and documenting output in notebook.
