### 2020-07-28 1e1ee09
Providing more info on output files in README and notebook.
