### 2020-07-29 #2436 @CFGrote
> Hi @CFGrote, based on your software's commit history it looks like you are the only author that made direct contributions to the code itself. Would you mind briefly stating what contributions the other authors listed in the paper made to the project?

All the commits are from me because I moved the code (initially developed by my co-authors) from a plain directory into github and then took care of integrating new features.

## Contributions by author:
- **Luka Opasic**: Development of the algorithm, initial version of the code, writing of manuscript
- **Jacob Scott** : contributed to the concept of the software and supervised its
development
- **Arne Traulsen**: supervised the development of the software.
- **Carsten Fortmann-Grote**: OO porting of the code, writing the manuscript, reference manual, documentation of code and examples, development of the test suite
