### 2020-07-29 4fb51ca
Adding output documentation to CancerSimulator.run()
