### 2020-07-29 mpievolbio-scicomp/cancer_sim#8 @CFGrote
> Similarly, for outputs generated by `casim.CancerSimulator.run`.

done (4fb51ca).
