### 2020-07-29 mpievolbio-scicomp/cancer_sim#16 @blakeaw
Software paper: Review
This issue is part of the JOSS review at https://github.com/openjournals/joss-reviews/issues/2436

Please note that this represents my initial review and does not take into account recent efforts by the submitting author to address issues.

# Software paper review

The work of Opasic et al. introduces the CancerSim software, which is a Python package used to simulate the somatic evolution of cancers. CancerSim employs a 2D lattice model, which allows the generational mutational profile and spatial heterogeneity of tumors to be modeled in an abstract manner as governed by key parameters related to cell division and mutation probabilities.

## Major Comments
- [ ] A detailed description of the simulation outputs is missing in both the paper and the software documentation (Documentation: Functional documentation). This may make it more difficult for users to fully evaluate if the software fits their needs (before using it), but perhaps more importantly, it is not initially clear how users can proceed with further analysis of the simulation outputs.
- [ ]  Although short descriptions of the key tumor models are provided in the paper¸ a detailed description of the key tumor model parameters and how they relate to each other is missing from both the paper and documentation (Documentation: Functional documentation).
- [ ] Although the summary of CancerSim itself is fairly detailed, there is no discussion of CancerSim in a broader context of cancer modeling methods or how CancerSim compares to other simulators. As such, I didn't get any sense of how CancerSim fits into the broader landscape of tumor modeling and cancer simulation methodologies. (Software paper: State of the field)

## Minor Comments

- [ ] This comment is related to the second major comment. There is no discussion or guidance on how users should determine key model parameters in a practical setting; e.g., should parameters be extracted directly from experimental measurements of cell division, or extracted from experimentally determined sequencing data (summary mentions reproducing sequencing data), or directly calibrated to some experimental data. I think some discussion and guidance in this area is needed to help users to evaluate the practical applicability of CancerSim to modeling their specific tumor systems, but will also be helpful for students and other non-specialists in further connecting the model to the tumor biology and available experimental methods.
- [ ] I think you may need to clarify that CancerSim installs as the `casim` package, both in the paper and the documentation.
- [ ] Summary, second paragraph, third sentence: This sentence is a little unclear to me. Does it mean that CancerSim can reproduce multi-region sequencing/profiling data? If so, I'd recommend re-wording a bit to make this more clear.
- [ ] Summary, second paragraph, fourth sentence: I'd recommend listing a few examples to help make the point more clearly.
- [ ]  Summary, first paragraph, third sentence: please change "existence tumour" to "existence a tumour".
- [ ] Summary, third paragraph, first sentence: I think a comma after "neoplasm type" is a little easier to read.
- [ ] Summary, third paragraph, second sentence: I'd suggest rewriting "It resembles the most to superficially" as "It most closely resembles superficially"
- [ ] Summary, fourth paragraph, sentence three: Please replace "mutation cell" with "mutation the cell".
- [ ] Summary, paragraph five, sentence three: Can the division probabilities change during the course of a simulation in response to a cell acquiring a specific mutation? Or are the division probabilities just fixed at the beginning in a way that models beneficial or deleterious effects?
