### 2020-07-29 #2436 @blakeaw
I have reviewed the CancerSim paper in https://github.com/mpievolbio-scicomp/cancer_sim/issues/16 and I grade the overall submission as Category 3) Major Revision.
Please note that this represents my initial evaluation and does not take into account recent efforts by @CFGrote to address some issues logged in the cancer_sim repository Issue tracker.
