### 2020-08-03 #2436 @pdebuyl
@jenzopr @blakeaw thank you for the reviews.

@CFGrote the bulk of the review so far concerns the paper and the documentation, can you propose a reply to the reviewers?

Regarding the paper, I will only point out that we request for JOSS to not include configuration files or tutorial-type instructions for running the code in the paper. We ask that this type of documentation remains in the README or the documentation of the code.
