### 2020-08-06 3760a0e
Documentation of output data.
