### 2020-08-06 e4161ff
Mention that cancer sim is provided as the 'casim'  module. (see review comments #16.)
