### 2020-08-06 b452789
Harmonized parameter names across code, tests, doc, paper. Add valid ranges in doc.
