+++ casim/casim.py
@@ -50,7 +50,7 @@ class CancerSimulatorParameters(object):
                  adv_mutant_death_probability=None,
                  mutation_probability=None,
                  adv_mutant_mutation_probability=None,
-                 mutations_per_division=None,
+                 number_of_mutations_per_division=None,
                  adv_mutation_interval=None,
                  number_of_initital_mutations=None,
                  tumour_multiplicity=None,
@@ -62,47 +62,47 @@ class CancerSimulatorParameters(object):
         """
         Construct a new CancerSimulationParameters object.
 
-        :param matrix_size: The size of the (square) grid in each dimension.
-        :type  matrix_size: int
+        :param matrix_size: The size of the (square) grid in each dimension. 
+        :type  matrix_size: int (matrix_size > 0)
 
         :param number_of_generations: The number of generations to simulate.
-        :type  number_of_generations: int
+        :type  number_of_generations: int (number_of_generations > 0)
 
         :param division_probability: The probability for a cell division to occur during one generation.
         :type  division_probability: float (0.0 <= division_probability <= 1.0)
 
         :param adv_mutant_division_probability: The probability for the division of a cell with advantageous mutation to occur during one generation.
-        :type  adv_mutant_division_probability: float (0.0 <= division_probability <= 1.0)
+        :type  adv_mutant_division_probability: float (0.0 <= adv_mutant_division_probability <= 1.0)
 
         :param death_probability: The probability for a cell to die during one generation.
-        :type  death_probability: float (0.0 <= division_probability <= 1.0)
+        :type  death_probability: float (0.0 <= death_probability <= 1.0)
 
         :param adv_mutant_death_probability: The probability for a cell with advantageous mutation to die during one generation.
-        :type  adv_mutant_death_probability: float (0.0 <= division_probability <= 1.0)
+        :type  adv_mutant_death_probability: float (0.0 <= adv_mutant_death_probability <= 1.0)
 
         :param mutation_probability: The probalitiy of mutation.
-        :type  mutation_probability: float (0.0 <= division_probability <= 1.0)
+        :type  mutation_probability: float (0.0 <= mutation_probability <= 1.0)
 
         :param adv_mutant_mutation_probability: The rate for an advantageous mutation to occur during one generation.
-        :type  adv_mutant_mutation_probability: float (0.0 <= division_probability <= 1.0)
+        :type  adv_mutant_mutation_probability: float (0.0 <= adv_mutant_mutation_probability <= 1.0)
 
-        :param mutations_per_division: The number of mutations per division
-        :type  mutations_per_division: int
+        :param number_of_mutations_per_division: The number of mutations per division
+        :type  number_of_mutations_per_division: int (0 < number_of_mutations_per_division)
 
         :param adv_mutation_interval: The number of generations after which an advantageous mutation can occur.
-        :type  adv_mutation_interval: int
+        :type  adv_mutation_interval: int (adv_mutation_interval > 0)
 
         :param number_of_initital_mutations: Number of mutations present in first cancer cell.
-        :type  number_of_initital_mutations: int
+        :type  number_of_initital_mutations: int (number_of_initital_mutations >= 0)
 
         :param tumour_multiplicity: Run in single or double tumour mode. Possible values: "single", "double".
         :type  tumour_multiplicity: str
 
-        :param read_depth: The sequencing depth (read length * number of reads / genome length). Default: 100.
-        :type  read_depth: int
+        :param read_depth: The sequencing read depth (read length * number of reads / genome length). Default: 100.
+        :type  read_depth: int (read_depth >= 0)
 
[…]
