+++ paper.md
@@ -36,29 +36,29 @@ proliferation, invasion, and destruction of the surrounding tissue
 [@Kumar2017]. Its high division and mutation rates
 lead to excessive intratumour genetic heterogeneity which makes cancer
 highly adaptable to environmental pressures such as therapy
-[@Turajlic2019]. Throughout most of its existence
-tumour is inaccessible to direct observation and experimental
-evaluation. Therefore, computational modelling can be useful to study
-many aspects of cancer. Some examples where theoretical models can be of
-great use include early carcinogenesis, as lesions are clinically
-observable when they already contain millions of cells, seeding of
-metastases, and cancer cell dormancy
-[@Altrock2015].
+[@Turajlic2019]. This process is known as somatic evolution of cancer.
+Throughout most of its existence a tumour is inaccessible to direct 
+observation and experimental evaluation. Therefore, computational modelling
+ can be useful to study many aspects of cancer. Some examples where theoretical 
+ models can be of great use include early carcinogenesis, as lesions are clinically
+observable when they already contain millions of cells, seeding of metastases, 
+and cancer cell dormancy [@Altrock2015].
 
 Here, we present CancerSim, a software that simulates somatic evolution of
 tumours. The software produces virtual spatial tumours with variable extent of
-intratumour genetic heterogeneity and realistic mutational profiles. Simulated
-tumours can be subjected to multi-region sampling to obtain mutation profiles
-that are realistic representation of the sequencing data. This makes the
-software useful for studying various sampling strategies in clinical cancer
-diagnostics. An early version of this cancer evolution model was used to
-simulate tumours subjected to sampling for classification of mutations based on
-their abundance [@Opasic2019]. Target users are scientists working in the field
-of mathematical oncology and students with interest in studying somatic
-evolution of cancer.  
+intratumour genetic heterogeneity and realistic mutational profiles. 
+Simulated tumours can be subjected to spatial sampling to obtain mutation profiles 
+from different tumour regions that are realistic representation of the sequencing data. 
+This makes the software useful for studying various sampling strategies in clinical cancer
+diagnostics such as needle biopsy sampling or liquid biopsy sampling. An early version of this 
+cancer evolution model was used to simulate tumours subjected to sampling for 
+classification of mutations based on their abundance [@Opasic2019]. Target users 
+are scientists working in the field of mathematical oncology. Simplicity of our model 
+in comparrisson to more advanced models like [@waclaw2015] makes it
+specifically suitable students with interest in studying somatic evolution of cancer.
 
 
-Our model is abstract, not specific to any neoplasm type and does not
+Our model is abstract, not specific to any neoplasm type, and does not
 consider a variety of biological features commonly found in neoplasm
 such as vasculature, immune contexture, availability of nutrients, and
 architecture of the tumour surroundings. It resembles the most to
@@ -79,13 +79,21 @@ every tumour cell in the tumour that has an unoccupied neighbour can
 divide with a certain probability (params.div\_\_probability). The
 daughter cell resulting from a cell division inherits all mutations from
 the parent cell and acquires a new mutation with a given probability
-(params.mut\_prob). Different division probabilities can be introduced
+(params.mut\_prob). Different division probabilities can be introduced in the beginning
 for some cells in order to simulate variability in fitness of cells that
 acquired a beneficial or deleterious mutation. The simulation allows the
 acquisition of more than one mutational event per cell
 (params.mut\_per\_division). In that case, variable amounts of
 sequencing noise [@Williams2016] can be added to make
-the output data more biologically realistic.
+the output data more biologically realistic. Key parameters params.number\_of\_generation, 
+params.division\_probability and params.death\_probability
[…]
