+++ casim/params.py
@@ -34,8 +34,8 @@ adv_mutant_mutation_probability = 1
 # Number of mutations per cell division (>=0).
 number_of_mutations_per_division = 10
 
-# Number of generation after which adv. mutations occur ([>=0]).
-adv_mutation_interval = 10
+# Number of generations after which adv. mutation occurs (>=0).
+adv_mutation_wait_time = 10
 
 # Number of mutations present in first cancer cell (>=0).
 number_of_initital_mutations = 150
