### 2020-08-07 9e2be3e
Improve wording, fix some bugs (parameters not taken over.)
