### 2020-08-10 2c31dda
Testing all parameters are correctly set when starting from params.py.
