+++ paper.md
@@ -57,7 +57,6 @@ are scientists working in the field of mathematical oncology. Simplicity of our 
 in comparrisson to more advanced models like [@waclaw2015] makes it
 specifically suitable students with interest in studying somatic evolution of cancer.
 
-
 Our model is abstract, not specific to any neoplasm type, and does not
 consider a variety of biological features commonly found in neoplasm
 such as vasculature, immune contexture, availability of nutrients, and
@@ -76,24 +75,23 @@ step.
 
 The simulation advances in discrete time-steps. In each simulation step,
 every tumour cell in the tumour that has an unoccupied neighbour can
-divide with a certain probability (params.div\_\_probability). The
+divide with a certain probability (params.division\_\_probability). The
 daughter cell resulting from a cell division inherits all mutations from
 the parent cell and acquires a new mutation with a given probability
-(params.mut\_prob). Different division probabilities can be introduced in the beginning
+(params.mutation\_probability). Different division probabilities can be introduced in the beginning
 for some cells in order to simulate variability in fitness of cells that
 acquired a beneficial or deleterious mutation. The simulation allows the
 acquisition of more than one mutational event per cell
-(params.mut\_per\_division). In that case, variable amounts of
+(params.number\_of\_mutations\_per\_division). In that case, variable amounts of
 sequencing noise [@Williams2016] can be added to make
-the output data more biologically realistic. Key parameters params.number\_of\_generation, 
+the output data more biologically realistic. Key parameters
+params.number\_of\_generations, 
 params.division\_probability and params.death\_probability
 determine the final size of the tumour, while the degree of intratumour heterogeneity can 
 be varied by changing the params.mutation\_probability parameter. 
 For neutral tumour evolution, parameter params.adv\_mutant\_division\_probability
 and params.adv\_mutant_death\_probability must be the same as params.division\_probability
-and params.death\_probability. Test
-
-
+and params.death\_probability. 
 
 Throughout the cancer growth phase, CancerSim stores information about
 the parent cell and a designation of newly acquired mutations for every
@@ -116,7 +114,6 @@ can be loaded from file and then subjected to the sampling process.
 
 Download and Installation
 -------------------------
-
 CancerSim is written in Python (version \>3.5). We recommend to install
 it directly from the source code hosted at github <https://github.com/mpievolbio-scicomp/cancer_sim>.
 
@@ -125,7 +122,6 @@ it directly from the source code hosted at github <https://github.com/mpievolbio
 
 Testing
 -------
-
 Although not strictly required, we recommend to run the test suite after
 installation. Simply execute the `run_tests.sh` shell script:
 
@@ -140,7 +136,7 @@ Results are published on
 
 High--level functionality
 -------------------------
-
+### Setting up the cancer simulation parameters
 The parameters of the cancer simulation are given via a python module or
 programmatically via the `CancerSimulationParameters` class. The file
 `params.py` is a documented parameter module:
@@ -177,16 +173,16 @@ number_of_mutations_per_division = 10
 adv_mutation_wait_time = 10
 
 # Number of mutations present in first cancer cell (>=0).
-number_of_initital_mutations = 150
+number_of_initial_mutations = 150
 
-# Tumour multiplicity ("single" || "double").
-tumour_multiplicity = "double"
+# Tumour multiplicity (one tumour or two tumours simultaneously) ("single" || "double").
+tumour_multiplicity = "single"
 
 # Sequencing read depth (read length * number of reads / genome length).
 read_depth = 100
 
 # Fraction of cells to be sampled ([0,1]).
-sampling_fraction = 0.9
[…]
