### 2020-08-10 c687d55
Testing all parameters are correctly set when starting from params.py.
