### 2020-08-10 6c8b664
Make sure log is also written to the log file in /log/casim.log.
