+++ casim/casim.py
@@ -25,16 +25,14 @@ from importlib.util import spec_from_file_location, module_from_spec
 
 np = numpy
 
-# Configure logging.
-LEVEL = logging.INFO
-LOGGER = logging.getLogger(__name__)
-LOGGER.setLevel(LEVEL)
-HANDLER = logging.StreamHandler()
-HANDLER.setLevel(LEVEL)
-FORMATTER=logging.Formatter('%(asctime)s %(levelname)s: %(message)s')
-HANDLER.setFormatter(FORMATTER)
-LOGGER.addHandler(HANDLER)
-
+LOG_FORMAT_STR = '%(asctime)s %(levelname)s: %(message)s'
+logging.basicConfig(
+    level=logging.INFO,
+    format=LOG_FORMAT_STR,
+    handlers=[
+        logging.StreamHandler(sys.stderr)
+    ]
+)
 
 class CancerSimulatorParameters(object):
     """
@@ -388,6 +386,13 @@ class CancerSimulator(object):
         self.__logdir = logdir
         self.__simdir = simdir
 
+        # Configure the logging filehandler.
+        root_logger = logging.getLogger()
+        fhandler = logging.FileHandler(os.path.join(logdir, "casim.log"))
+        fhandler.setFormatter(root_logger.handlers[0].formatter)
+        import ipdb; ipdb.set_trace()
+        root_logger.addHandler(fhandler)
+
     def extend_sample(self, sample_center, sample_size):
         """ Takes a subset of cells from the tumour positioned around single input cell with specific coordinates. Output is a list of tuples of cells belonging to the sample.
         :param sample_center: coordinates of cell that will be center of the sample
@@ -471,10 +476,10 @@ class CancerSimulator(object):
         matrix_size=self.parameters.matrix_size
 
         if self.__tumour_multiplicity == 'single':
-            LOGGER.info('Running in single tumour mode.')
+            logging.info('Running in single tumour mode.')
             initLoc=(int(matrix_size*0.5),int(matrix_size*0.5))
 
-            LOGGER.info("First cell at %s.", str(initLoc))
+            logging.info("First cell at %s.", str(initLoc))
 
             self.__mtx[initLoc]=1
             self.__mut_container=[(0, 0), (0, 1)]
@@ -482,7 +487,7 @@ class CancerSimulator(object):
 
         #start the pool of cancer cells by adding the initial cancer cell into it
         if self.__tumour_multiplicity == 'double':
-            LOGGER.info('Running in sdsa mode.')
+            logging.info('Running in sdsa mode.')
 
             ### COMMENT: Should these be given as parameters?
             distance_between_tumours=0.05
@@ -499,7 +504,7 @@ class CancerSimulator(object):
         self.__growth_plot_data=[]
 
         self.__pre_run_log()
-        LOGGER.info('Tumour growth in progress.')
+        logging.info('Tumour growth in progress.')
 
         death_list=[]
         prop_of_driver=[]
@@ -557,21 +562,21 @@ class CancerSimulator(object):
 
         self.__post_run_log()
 
-        LOGGER.info("Consumed Wall time of this run: %f s.", end - start)
+        logging.info("Consumed Wall time of this run: %f s.", end - start)
 
         return 0
 
     def __pre_run_log(self):
         message = ""
 
-        LOGGER.info("Ready to start CancerSim run with these parameters:")
+        logging.info("Ready to start CancerSim run with these parameters:")
         for k,v in self.parameters.__dict__.items():
-            LOGGER.info("%s = %s", k.split("__")[-1], v)
+            logging.info("%s = %s", k.split("__")[-1], v)
     
     def __post_run_log(self):
[…]
