### 2020-08-10 dea1333
Make sure log is also written to the log file in /log/casim.log.
