### 2020-08-10 16c2d5c
Attempt to render citations as hyperlinks.
