### 2020-08-10 7f42e9d
Attempt to render citations as hyperlinks.
