### 2020-08-10 9d3b4e8
Attempt to render citations as hyperlinks.
