### 2020-08-11 #2436 @CFGrote
sorry for not being specific:
i meant the checkboxes in the lists of review comments in  https://github.com/mpievolbio-scicomp/cancer_sim/issues/16 and  https://github.com/mpievolbio-scicomp/cancer_sim/issues/11 .
