### 2020-08-11 29aa12e
Some fixes and typos in doc and paper. Removed tutorial-like code and file listings from the paper as demanded by @pdebuyl in https://github.com/openjournals/joss-reviews/issues/2436#issuecomment-667910281.
