+++ paper.md
@@ -44,7 +44,7 @@ observation and experimental evaluation. Therefore, computational modelling
 observable when they already contain millions of cells, seeding of metastases, 
 and cancer cell dormancy [@Altrock2015].
 
-Here, we present CancerSim, a software that simulates somatic evolution of
+Here, we present `CancerSim`, a software that simulates somatic evolution of
 tumours. The software produces virtual spatial tumours with variable extent of
 intratumour genetic heterogeneity and realistic mutational profiles. 
 Simulated tumours can be subjected to spatial sampling to obtain mutation profiles 
@@ -52,14 +52,14 @@ from different tumour regions that are realistic representation of the sequencin
 This makes the software useful for studying various sampling strategies in clinical cancer
 diagnostics such as needle biopsy sampling or liquid biopsy sampling. An early version of this 
 cancer evolution model was used to simulate tumours subjected to sampling for 
-classification of mutations based on their abundance [@Opasic2019]. Target users 
-are scientists working in the field of mathematical oncology. Simplicity of our model 
-in comparison to more advanced models like [@waclaw2015] makes it
+classification of mutations based on their abundance [@Opasic2019]. Target users
+of `CancerSim` are scientists working in the field of mathematical oncology. Simplicity of our model 
+in comparison to more advanced models (see e.g. Ref. [@waclaw2015]) makes it
 specifically suitable for students with interest in somatic evolution of cancer.
 
 Our model is abstract, not specific to any neoplasm type, and does not
 consider a variety of biological features commonly found in neoplasm
-such as vasculature, immune contexture, availability of nutrients, and
+such as blood vessels, immune cells, availability of nutrients, and
 architecture of the tumour surroundings. It resembles the most to
 superficially spreading tumours like carcinoma in situ, skin cancers, or
 gastric cancers, but it can be used to model any tumour on this abstract
@@ -93,7 +93,7 @@ For neutral tumour evolution, parameter `adv_mutant_division_probability`
 and `adv_mutant_death_probability` must be the same as `division_probability`
 and `death_probability`. 
 
-Throughout the cancer growth phase, CancerSim stores information about
+Throughout the cancer growth phase, `CancerSim` stores information about
 the parent cell and a designation of newly acquired mutations for every
 cell. Complete mutational profiles of cells are reconstructed a
 posteriori based on the stored lineage information.
@@ -102,7 +102,8 @@ The division rules which allow only cells with empty neighbouring nodes
 to divide, cause exclusively peripheral growth and complete absence of
 dynamics in the tumour centre. To allow for variable degree of growth
 inside the tumour, we introduced a death process. At every time step,
-after all cells attempt their division, a number of random cells die and
+after all cells attempt their division, a number of random cells die according
+to `death_probability` and `adv_mutant_death_probability` and
 yield their position to host a new cancer cell in a subsequent time
 step.
 
@@ -114,11 +115,13 @@ can be loaded from file and then subjected to the sampling process.
 
 Download and Installation
 -------------------------
-CancerSim is written in Python (version \>3.5). We recommend to install
+`CancerSim` is written in Python (version \>3.5). We recommend to install
 it directly from the source code hosted at github <https://github.com/mpievolbio-scicomp/cancer_sim>.
 
  Detailed instructions including creation of a
-`conda` environment are given in the online documentation at <https://cancer-sim.readthedocs.io/en/master/include/README.html#installation>.
+`conda` environment are given in the online documentation at
+<https://cancer-sim.readthedocs.io/en/master/include/README.html#installation>.
[…]
