### 2020-08-11 70422f7
Adding proper caption to figure.
