### 2020-08-11 453e29e
Set graphic size.
