### 2020-08-11 e940122
Revert "Set graphic size."
This reverts commit 453e29e75cbfdb3c1d5942b00262935716742662.
