### 2020-08-11 #2436 @pdebuyl
@CFGrote even though these are issues in your repository, I think that it makes sense for the reviewers to check them. They will go through their remarks anyway while assessing the changes.
