### 2020-08-12 8c1f7d5
addressing minor comments from review #11
