+++ README.md
@@ -37,9 +37,9 @@ this cancer evolution model was used to simulate tumours subjected to
 sampling for classification of mutations based on their abundance
 \[[4](#ref-opasic2019)\].
 Target users of `CancerSim` are scientists working in the field of
-mathematical oncology. Simplicity of our model in comparison to more advanced
-models (see e.g.  Ref. \[[5](#ref-waclaw2015)\])
-makes it specifically suitable for students with interest in somatic evolution of cancer.
+mathematical oncology. Simplicity and accessibility of our model in comparison to more advanced
+models (see e.g.  Ref. \[[5](#ref-waclaw2015)\]) makes it particularly 
+suitable for students with interest in somatic evolution of cancer.
 
 Our model is abstract, not specific to any neoplasm type, and does not
 consider a variety of biological features commonly found in neoplasm
@@ -62,17 +62,17 @@ every tumour cell in the tumour that has an unoccupied neighbour can
 divide with a certain probability (set by the parameter `division_probability`). The
 daughter cell resulting from a cell division inherits all mutations from
 the parent cell and acquires a new mutation with a given probability
-`mutation_probability`). Different division probabilities can be introduced in the beginning
-for some cells in order to simulate variability in fitness of cells that
-acquired a beneficial or deleterious mutation. The simulation allows the
-acquisition of more than one mutational event per cell
+`mutation_probability`). A new mutation that changes death and birth probability of cell can be introduced
+at into random cell at the specific time step defined by `adv_mutation_wait_time`.
+By changing fitness parameters of a mutant cell `adv_mutant_division_probability`
+and `adv_mutant_death_probability` one can model various evolutionary processes
+like emergence of a faster dividing sub-clone or selective effects of a drug treatment.
+
+The simulation allows the acquisition of more than one mutational event per cell
 (`number_of_mutations_per_division`). In that case, variable amounts of
-sequencing noise
-\[[6](#ref-williams2016)\]
-can be added to make
+sequencing noise \[[6](#ref-williams2016)\] can be added to make
 the output data more biologically realistic. The key parameters
-`number_of_generations`, 
-`division_probability` and `death_probability`
+`number_of_generations`, `division_probability` and `death_probability`
 determine the final size of the tumour, while the degree of intratumour heterogeneity can 
 be varied by changing the `mutation_probability` parameter. 
 For neutral tumour evolution, parameter `adv_mutant_division_probability`
