### 2020-08-13 93e47a0
Implement and document restart from dump.
