+++ test/casim_test.py
@@ -2,7 +2,7 @@
 
 # Import class to be tested.
 from casim import casim
-from casim.casim import CancerSimulator, CancerSimulatorParameters, check_set_number, load_cancer_simulation, LOGGER
+from casim.casim import CancerSimulator, CancerSimulatorParameters, check_set_number, load_cancer_simulation
 
 from collections import namedtuple
 from test_utilities import _remove_test_files
@@ -17,6 +17,8 @@ import unittest
 import pickle
 import pandas
 from tempfile import mkdtemp
+
+logging.getLogger().setLevel(logging.DEBUG)
 
 
 class CancerSimulatorParametersTest(unittest.TestCase):
@@ -199,7 +201,7 @@ class CancerSimulatorTest(unittest.TestCase):
 
         # Capture stdout.
         stream = StringIO()
-        log = LOGGER
+        log = logging.getLogger()
         for handler in log.handlers:
            log.removeHandler(handler)
         myhandler = logging.StreamHandler(stream)
@@ -401,37 +403,36 @@ class CancerSimulatorTest(unittest.TestCase):
     def test_serialize(self):
         """ The the serialization of the entire object. """
         parameters = CancerSimulatorParameters()
-        cancer_sim = CancerSimulator(parameters, seed=1, outdir=mkdtemp())
+        cancer_sim = CancerSimulator(parameters, seed=2, outdir=mkdtemp())
 
         # Cleanup.
         self._test_files.append(cancer_sim.outdir)
 
-        # dump before run.
+        # Dump.
         cancer_sim.dump()
 
         # Reload
         loaded_simulation = load_cancer_simulation(cancer_sim.dumpfile)
-
         self.assertIsInstance(loaded_simulation, CancerSimulator)
 
         # Check parameters.
         loaded_parameters = loaded_simulation.parameters
 
-        self.assertEqual(loaded_parameters.number_of_generations,                parameters.number_of_generations)
-        self.assertEqual(loaded_parameters.matrix_size,                          parameters.matrix_size)
-        self.assertEqual(loaded_parameters.number_of_generations,                parameters.number_of_generations)
-        self.assertEqual(loaded_parameters.division_probability,                 parameters.division_probability)
-        self.assertEqual(loaded_parameters.adv_mutant_division_probability,    parameters.adv_mutant_division_probability)
-        self.assertEqual(loaded_parameters.death_probability,                    parameters.death_probability)
-        self.assertEqual(loaded_parameters.adv_mutant_death_probability,  parameters.adv_mutant_death_probability)
-        self.assertEqual(loaded_parameters.mutation_probability,                        parameters.mutation_probability)
-        self.assertEqual(loaded_parameters.adv_mutant_mutation_probability,    parameters.adv_mutant_mutation_probability)
-        self.assertEqual(loaded_parameters.number_of_mutations_per_division,               parameters.number_of_mutations_per_division)
-        self.assertEqual(loaded_parameters.adv_mutation_wait_time,        parameters.adv_mutation_wait_time)
-        self.assertEqual(loaded_parameters.number_of_initial_mutations,                     parameters.number_of_initial_mutations)
-        self.assertEqual(loaded_parameters.tumour_multiplicity,              parameters.tumour_multiplicity)
-        self.assertEqual(loaded_parameters.read_depth,              parameters.read_depth)
-        self.assertEqual(loaded_parameters.sampling_fraction,              parameters.sampling_fraction)
+        self.assertEqual(loaded_parameters.number_of_generations, parameters.number_of_generations)
[…]
