+++ casim/casim.py
@@ -129,6 +129,7 @@ class CancerSimulatorParameters(object):
         self.plot_tumour_growth = plot_tumour_growth
         self.export_tumour = export_tumour
     
+        
     @property
     def matrix_size(self):
         return self.__matrix_size
@@ -293,24 +294,54 @@ class CancerSimulator(object):
         self.parameters = parameters
 
         # Setup internal variables.
-        self.__prop_of_driver = None
         self.__mtx = lil_matrix((self.parameters.matrix_size, self.parameters.matrix_size), dtype=int)
         self.__mut_container = None
-        self.__lq_bipsy = None
         self.__xaxis_histogram = None
         self.__death_list = None
         self.__biopsy_timing = None
         self.__beneficial_mutation = []
         self.__growth_plot_data = None
-        self.__mutation_counter = None
         self.__s = self.parameters.number_of_mutations_per_division
-        self.__tumour_multiplicity = self.parameters.tumour_multiplicity
+
+        # Keep track of how many steps where performed in previous run if this
+        # is a reloaded run.
+        self.__init_step = 0
+        # Keep track of mutation count in previous run if this is a rerun.
+        self.__mutation_counter = 1
 
         # Handle direct parameters.
         self.seed = seed
         self.outdir = outdir
         self.__ploidy=2
         self.__mut_multiplier=[self.__s]*100000
+        
+        if self.parameters.tumour_multiplicity == 'single':
+            logging.info('Running in single tumour mode.')
+            initLoc=(int(self.parameters.matrix_size*0.5),int(self.parameters.matrix_size*0.5))
+
+            logging.info("First cell at %s.", str(initLoc))
+
+            self.__mtx[initLoc]=1
+            self.__mut_container=[(0, 0), (0, 1)]
+            self.__pool=[initLoc]
+
+        #start the pool of cancer cells by adding the initial cancer cell into it
+        if self.parameters.tumour_multiplicity == 'double':
+            logging.info('Running in sdsa mode.')
+
+            ### COMMENT: Should these be given as parameters?
+            distance_between_tumours=0.05
+            initLoc=(int(self.parameters.matrix_size*0.45),int(self.parameters.matrix_size*0.5))
+            secondinitLoc=(int(self.parameters.matrix_size*0.65),int(self.parameters.matrix_size*0.51))
+
+            self.__mtx[initLoc]=1
+            self.__mtx[secondinitLoc]=2
+            self.__mut_container=[(0, 0), (0, 1), (0,2)]
+            self.__pool=[initLoc, secondinitLoc]
+
+        # create lists used in loops
+        self.__growth_plot_data=[]
+
 
     @property
     def seed(self):
@@ -474,39 +505,8 @@ class CancerSimulator(object):
         # Setup square matrix.
         matrix_size=self.parameters.matrix_size
 
-        if self.__tumour_multiplicity == 'single':
-            logging.info('Running in single tumour mode.')
-            initLoc=(int(matrix_size*0.5),int(matrix_size*0.5))
-
-            logging.info("First cell at %s.", str(initLoc))
-
-            self.__mtx[initLoc]=1
-            self.__mut_container=[(0, 0), (0, 1)]
-            self.__pool=[initLoc]
-
-        #start the pool of cancer cells by adding the initial cancer cell into it
-        if self.__tumour_multiplicity == 'double':
-            logging.info('Running in sdsa mode.')
-
-            ### COMMENT: Should these be given as parameters?
-            distance_between_tumours=0.05
-            initLoc=(int(matrix_size*0.45),int(matrix_size*0.5))
-            secondinitLoc=(int(matrix_size*0.65),int(matrix_size*0.51))
-
[…]
