### 2020-08-19 0d9d131
Use numpy.random everywhere, remove Lib.random
