### 2020-08-19 02e2112
Revert "Use numpy.random everywhere, remove Lib.random"
This reverts commit 0d9d131140a1aa05a471a22b41bd826395c209b3.
