+++ casim/casim.py
@@ -7,6 +7,7 @@ __version__ = '0.0.1'
 
 from argparse import ArgumentParser
 from operator import itemgetter
+from random import shuffle
 from scipy.sparse import lil_matrix
 from time import sleep, time
 from timeit import default_timer as timer
@@ -18,7 +19,7 @@ import matplotlib.pyplot as plt
 import numpy
 import os
 import pickle
-from numpy import random as prng
+import random as prng
 import sys
 from importlib.util import spec_from_file_location, module_from_spec
 
@@ -360,7 +361,6 @@ class CancerSimulator(object):
         if not val > 0:
             raise ValueError("The parameter 'seed' must a positive integer (int).")
 
-        prng.seed(val)
         self.__seed = val
 
     @property
@@ -510,6 +510,9 @@ class CancerSimulator(object):
 
         start = timer()
 
+        seed=self.__seed
+        prng.seed(seed)
+
         #run growth function
         #output variable (true_vaf) is list of tuples with mutation id and frequency of mutation in the tumour [(mut_id, frequency),...]
         true_vaf=self.tumour_growth()
@@ -522,7 +525,7 @@ class CancerSimulator(object):
         # Setup list of coordinates that serve as center of sampling [(x,y)]
         # Pick a random cell from the pool.
         # TODO: enter wanted sampling coordinates.
-        random_index = prng.randint(len(self.__pool))
+        random_index = numpy.random.randint(len(self.__pool))
         samples_coordinates_list=[self.__pool[random_index]]
 
         #iterate over each sample from the list of samples
@@ -583,6 +586,9 @@ class CancerSimulator(object):
         :param sample_coordinates: coordinates of central sample cell
         :type  sample_coordinates: tuple (i,j) of cell indices
         """
+
+
+
         xaxis_histogram=np.arange(0.0,1,0.01)
         #setdetection limit of the mutation in the sample (depends on the sequencing machine and sequencing depth)
         detection_limit=0.05
@@ -756,13 +762,13 @@ class CancerSimulator(object):
 
         return extended_mut_list
 
-    def terminate_cell(self, cell, t_step):
+    def terminate_cell(self, cell, step):
         """ Kills cancer cell and removes it from the pool of cancer cells
 
         :param cell: cell chosen for termination
         :type cell: tuple (i,j) of cell indices.
 
-        :param int t_step: The time step in the simulation
+        :param int step: The time step in the simulation
         """
         #removes cell from the pool
         self.__pool.remove(cell)
@@ -770,17 +776,16 @@ class CancerSimulator(object):
         #resets value of position on matrix to zero
         self.__mtx[cell]=0
 
-    def death_step(self, t_step):
+    def death_step(self, step):
 
         """ Takes a group of random cells and kills them
 
-        :param int t_step: The time step in the simulation
+        :param int step: The time step in the simulation
 
         """
-        for cell in self.__pool:
-            r = prng.random()
-            if (cell in self.__beneficial_mutation and r < self.parameters.adv_mutant_death_probability) or r < self.parameters.death_probability:
-                self.terminate_cell(cell, t_step)
+        for i in prng.sample(self.__pool, math.floor(self.parameters.death_probability*len(self.__pool))):
+            self.terminate_cell(i, step)
+
 
     def mutation_reconstruction(self,cells_to_reconstruct):
         """ Reconstructs list of mutations of individual cell by going thorough its ancestors.
@@ -836,15 +841,15 @@ class CancerSimulator(object):
             mutation_counter = self.__mutation_counter + 1
 
         # Loop over time steps.
-        for t_step in range(self.__init_step, self.__init_step+self.parameters.number_of_generations):
[…]
