### 2020-08-19 ed20551
Cleanup conditional statements and logic.
