### 2020-08-19 24d97c0
Log beneficial events. Notebook updated."
