### 2020-08-19 7e32183
Implemented sample position user interface. This disturbed the sequence of random numbers, hence had to make new reference data.
