+++ test/casim_test.py
@@ -66,6 +66,7 @@ class CancerSimulatorParametersTest(unittest.TestCase):
         self.assertEqual(parameters.tumour_multiplicity, 'single')
         self.assertEqual(parameters.read_depth, 100)
         self.assertEqual(parameters.sampling_fraction, 0.0)
+        self.assertIsNone(parameters.sampling_positions)
         self.assertTrue(parameters.plot_tumour_growth)
         self.assertTrue(parameters.export_tumour)
 
@@ -87,6 +88,7 @@ class CancerSimulatorParametersTest(unittest.TestCase):
                                 tumour_multiplicity='single',
                                 read_depth=200,
                                 sampling_fraction=0.3,
+                                sampling_positions=[(13,12), (3,6)],
                                 export_tumour=False,
                                 plot_tumour_growth=False,
                                                 )
@@ -105,6 +107,7 @@ class CancerSimulatorParametersTest(unittest.TestCase):
         self.assertEqual(parameters.tumour_multiplicity, 'single' )
         self.assertEqual(parameters.read_depth, 200 )
         self.assertEqual(parameters.sampling_fraction, 0.3 )
+        self.assertEqual(numpy.linalg.norm(parameters.sampling_positions - numpy.array([[13,12], [3,6]])), 0)
         self.assertFalse(parameters.plot_tumour_growth)
         self.assertFalse(parameters.export_tumour)
 
@@ -328,8 +331,6 @@ class CancerSimulatorTest(unittest.TestCase):
         ### Load results and reference data.
         # Reference data.
         ref_dir = os.path.join(os.path.abspath(os.path.dirname(__file__)), 'reference_test_data_50mut', 'cancer_1', 'simOutput')
-        with open(os.path.join(ref_dir, 'death_list.p'), 'rb') as fp:
-            ref_death_list = pickle.load(fp)
         with open(os.path.join(ref_dir, 'mtx.p'), 'rb') as fp:
             ref_mtx = pickle.load(fp)
         with open(os.path.join(ref_dir, 'mut_container.p'), 'rb') as fp:
@@ -338,15 +339,12 @@ class CancerSimulatorTest(unittest.TestCase):
 
         run_dir = os.path.join('reference_test_out', 'cancer_1', 'simOutput')
         # Run data.
-        with open(os.path.join(run_dir, 'death_list.p'), 'rb') as fp:
-            run_death_list = pickle.load(fp)
         with open(os.path.join(run_dir, 'mtx.p'), 'rb') as fp:
             run_mtx = pickle.load(fp)
         with open(os.path.join(run_dir, 'mut_container.p'), 'rb') as fp:
             run_mutations = pickle.load(fp)
 
         # Check data is equal.
-        self.assertEqual(ref_death_list, run_death_list)
         self.assertAlmostEqual(numpy.linalg.norm((ref_mtx - run_mtx).toarray()), 0.0)
         self.assertEqual(ref_mutations, run_mutations)
 
@@ -369,6 +367,7 @@ class CancerSimulatorTest(unittest.TestCase):
                                             number_of_mutations_per_division=1,
                                             tumour_multiplicity=None,
                                             sampling_fraction=0.1,
+                                            sampling_positions=[(501,502)],
                                             )
 
         simulator = CancerSimulator(parameters=parameters, seed=1, outdir="reference_test_out")
@@ -379,8 +378,6 @@ class CancerSimulatorTest(unittest.TestCase):
         ### Load results and reference data.
         # Reference data.
         ref_dir = os.path.join(os.path.abspath(os.path.dirname(__file__)), 'reference_test_data_1mut', 'cancer_1', 'simOutput')
-        with open(os.path.join(ref_dir, 'death_list.p'), 'rb') as fp:
-            ref_death_list = pickle.load(fp)
[…]
