+++ casim/casim.py
@@ -2,7 +2,7 @@
 #!/usr/bin/env python3
 __author__ = 'Luka Opasic, MD'
 __email__ = 'opasic@evolbio.mpg.de'
-__version__ = '0.0.1'
+__version__ = '1.1.0'
 
 
 from argparse import ArgumentParser
@@ -56,6 +56,7 @@ class CancerSimulatorParameters(object):
                  sampling_fraction=None,
                  plot_tumour_growth=None,
                  export_tumour=None,
+                 sampling_positions=None,
                 ):
         """
         Construct a new CancerSimulationParameters object.
@@ -103,6 +104,9 @@ class CancerSimulatorParameters(object):
         :param sampling_fraction: The fraction of cells to include in a sample. Default: 0.
         :type  sampling_fraction: float  (0 <= sampling_fraction <= 1)
 
+        :param sampling_positions: The positions of cells to include in a sample. Default: Random position.
+        :type  sampling_positions: List (or array) of tuples of ints. E.g.  ([10,20], [2,31]).
+
         :param plot_tumour_growth: Render graph of the tumour size as function
         of time. Default: True.
         :type plot_tumour_growth: bool
@@ -126,6 +130,7 @@ class CancerSimulatorParameters(object):
         self.tumour_multiplicity = tumour_multiplicity
         self.read_depth = read_depth
         self.sampling_fraction = sampling_fraction
+        self.sampling_positions = sampling_positions
         self.plot_tumour_growth = plot_tumour_growth
         self.export_tumour = export_tumour
     
@@ -238,6 +243,26 @@ class CancerSimulatorParameters(object):
         self.__sampling_fraction = check_set_number(val, float, 0.0, 0.0, 1.0)
 
     @property
+    def sampling_positions(self):
+        return self.__sampling_positions
+    @sampling_positions.setter
+    def sampling_positions(self, val):
+        if val is not None:
+            for pos in val:
+                if not hasattr(val, "__iter__"):
+                    raise TypeError("Sampling positions must be list of tuples.")
+                if len(pos) != 2:
+                    raise ValueError("Sampling positions must be list of 2-tuples (x,y coordinates).")
+                for xy in pos:
+                    if not isinstance(xy, int):
+                        raise TypeError("Sampling position must be integer")
+                    if xy < 0 or xy > self.matrix_size:
+                        raise ValueError("Sampling position must be positive integer not larger than the matrix size.")
+
+        self.__sampling_positions = val
+            
+
+    @property
     def plot_tumour_growth(self):
         return self.__plot_tumour_growth
     @plot_tumour_growth.setter
@@ -297,11 +322,11 @@ class CancerSimulator(object):
         self.__mtx = lil_matrix((self.parameters.matrix_size, self.parameters.matrix_size), dtype=int)
         self.__mut_container = None
         self.__xaxis_histogram = None
-        self.__death_list = None
         self.__biopsy_timing = None
         self.__beneficial_mutation = []
         self.__growth_plot_data = None
         self.__s = self.parameters.number_of_mutations_per_division
+
 
         # Keep track of how many steps where performed in previous run if this
         # is a reloaded run.
@@ -434,32 +459,34 @@ class CancerSimulator(object):
 
         biopsy_size=math.ceil(sample_size*len(self.__pool))
 
-        if biopsy_size>1:
+        #look at z-tier neighbours around sample_center
+        for z in range(1,len(self.__pool)):
+            expanded_sample=[]
 
             #look at z-tier neighbours around sample_center
-            for z in range(1,len(self.__pool)):
[…]
