### 2020-08-19 cb8e293
Fixing remaining tests.
