### 2020-08-19 839c23a
Update doc and params.py. Fix to main() to pick up new parameter 'sampling_positions'
