### 2020-08-20 4581588
Clarified multiregion sampling in paper.
