### 2020-08-20 95a30d4
Moving discussion of future directions to Summary.
