### 2020-08-28 e05b3f8
Removed unneeded rename from installation instructions.
