### 2020-08-31 #2436 @CFGrote
I have just submitted our responses to the reviews: https://github.com/mpievolbio-scicomp/cancer_sim/issues/11#issuecomment-683706999 and https://github.com/mpievolbio-scicomp/cancer_sim/issues/16#issuecomment-683707010 All relevant changes are pushed to the master branch in https://github.com/mpievolbio-scicomp/cancer_sim.

Thanks again to @blakeaw  and @jenzopr for the careful reading and detailed comments.
