### 2020-08-31 305492d
Delete response_to_referees.md
