+++ response_to_referees.md
@@ -1,194 +1,0 @@
-# Reply to software paper review issue #11
-
-
-> This issue is part of the JOSS review at [openjournals/joss-reviews#2436](https://github.com/openjournals/joss-reviews/issues/2436)
-> ## Software paper review
-> 
-> The work of Opasic and colleagues introduces CancerSim, a python package to study the evolution of somatic mutations in cancer cells, by introducing a simple simulation procedure on a 2D lattice and along discrete time steps. The package features key parameters, like cell division and mutation probabilities, tracks parent-daugther cell relationships and allows for adjustment of cell fitness of mutated cells. It therefore resembles an abstract model of superficially spreading tumors.
-> 
-> Although the paper consists of a well-written summary and an in-depth algorithm outline, description of key parameters remains superficial and applicability of the package in theoretical modelling is only vaguely delineated.
-> ### Major comments
-> 
->     * [ ]  The second paragraph of the **Summary** dives straight into specialist concepts like "cancer evolution", "somatic mutations", "genetic heterogeneity" and "sequencing". To enhance understanding for a diverse, non-specialist audience, please enhance the short conceptual introduction in the first paragraph, picking up those concepts. (Software paper: Summary)
-We explained or replaced the technical/scientific terms by more general language
-to enhance the accessibility.
-Relevant commits: 9e2be3e
-> 
->     * [ ]  There exists some discrepancy between parameter description and parameter naming (e.g. the number of divisions per generation vs. `div_probability`, number of division for cells with mutation vs. `fittnes_advantage_div_prob`). Please make sure to revise parameter names to indicate e.g. if a probability or absolute number is required, and include descriptions of which and how parameters interplay or depend on each other. (Documentation: Functional documentation)
-All parameter names have been made consistent across the source code, the API
-reference manual, README, paper, and online documentation. We explain the role
-of all parameters in the template parameter module `params.py` by commenting on the respective line.
- This file is
-reproduced in the README, followed by additional explanations of each parameter
-and their interplay. The `CancerSimulationParameters` API reference manual also
-explains all parameters. A table with all parameter names, their function and
-permitted values is added to the paper, replacing the code listing of the
-`params.py` template parameter module. Relevant commits:
-d6de50, b45278, 4a39b4d.
-> 
->     * [ ]  A description of simulation output is missing in the paper and the online help. Having a detailed description of the output files lets readers quickly judge wheter or not a software will be suitable for their needs. (Documentation: Functional documentation)
-We have added documentation of the simulation outputs to the paper, the README
-and the API documentation (`CancerSimulator.run()` method). Also the quickstart
-example now explains the output and renders the produced figures (mutation
-profiles and growth curve.) Related commits:
-3760a0, 4fb51c, 1e1ee0, 61220b .
-> 
->     * [ ]  Having a detailed description of simulation output will also help readers to envision how the software can be applied in theoretical modelling and, importantly, which related research questions can be addressed e.g. in follow-up studies. Please delineate the softwares potential in much more detail, by suggestions or elaboration on possible downstream analysis. (Software paper: Statement of need). 
-The word limit in JOSS allows us only to give a short discussion on potential
-applications of our software. This has been added to the paper (Summary, last
-paragraph.
-Relevant commit: 5187dee, 5753fb5
-> 
-> 
-> ### Minor comments
-> 
->     * [ ]  The first paragraph mentions cancer cell dormancy as an example for beneficial use of theoretical models to understand carcinogenesis. Unfortunately, CancerSim does not allow cells to be simulated dormant, which renders the example detrimental to the papers statement.
-Simulation of dormant cancer cells requires changing the growth parameters, in
[…]
