### 2020-09-01 mpievolbio-scicomp/cancer_sim#11 @jenzopr
Thanks a lot for the detailed response. I consider my remarks fully addressed. Well done 👍
