### 2020-09-02 #2436 @pdebuyl
@blakeaw the reply to your review is at https://github.com/mpievolbio-scicomp/cancer_sim/issues/16
