### 2020-09-06 mpievolbio-scicomp/cancer_sim#16 @blakeaw
Thanks for the detailed responses. I consider my comments here to be appropriately addressed.
