### 2020-09-06 mpievolbio-scicomp/cancer_sim#10 @blakeaw
Thanks for addressing this issue. I'll go ahead and close.
For future reference, the Quickstart example and its description are in the documentation here: https://cancer-sim.readthedocs.io/en/master/include/notebooks/quickstart_example.html
