### 2020-09-06 mpievolbio-scicomp/cancer_sim#9 @blakeaw
Ok, thanks for checking.  Everything shows up correctly for me now. There must have been an issue on my end earlier causing it not to load properly. Closing.
