### 2020-09-06 mpievolbio-scicomp/cancer_sim#8 @blakeaw
Thanks! Since this issue has been addressed I'll go ahead an close.
