### 2020-09-09 #2436 @pdebuyl
I am not subscribed to the issues at the repo and did not see @blakeaw 's positive evaluation of the changes at https://github.com/mpievolbio-scicomp/cancer_sim/issues/16 . @CFGrote I will proceed to the coming editorial steps in the coming days.
