### 2020-09-09 #2436 @blakeaw
Sorry about that @pdebuyl. I forgot to come back and comment here after I finished reviewing the changes.
I've checked off all the boxes in my Reviewer Checklist except the Functionality:Functionality one. For that, I'm planning to execute a few more test simulations with different parameter sets to make sure the code operates smoothly outside of the provided template/examples. I'll update within the next day or two.
Also, it looks like @jenzopr still had an issue with installation and unit tests of the updated code: recorded in https://github.com/mpievolbio-scicomp/cancer_sim/issues/6.
