### 2020-09-14 mpievolbio-scicomp/cancer_sim#6 @pdebuyl
Hi,

I can run the tests properly. Looking at the line numbers of the errors above it is likely that an old version of the package remains. Could you try to remove (manually by going the the site-package directory and `rm -r casim*`) and reinstall the package?
