### 2020-09-14 #2436 @pdebuyl
@CFGrote there still is a testing issue open at your software's repository. I could run the tests just file (python 3.7 linux amd64 using pip), so I guess that the issue can be resolved by updating completely the software on the user's machine.
