### 2020-09-14 mpievolbio-scicomp/cancer_sim#6 @CFGrote
@jenzopr , could you please run the tests again *after* purging all casim* files from your environment and reinstalling, as per @pdebuyl's comment above?

thanks!
