### 2020-09-14 mpievolbio-scicomp/cancer_sim#6 @jenzopr
Purging the env and reinstalling from scratch worked. Thanks!
