### 2020-09-14 #2436 @blakeaw
Hi @pdebuyl, Functionality doesn't require any additional work. You can consider it checked.
(The checkbox is showing up as checked on my end. Maybe refreshing the page will update?)
