### 2020-09-17 a612ec1
Update setup.py
Bump version to 2.0.1
