+++ setup.py
@@ -2,7 +2,7 @@ from setuptools import setup
 
 setup(
     name='CaSim',
-    version='1.0.0',
+    version='2.0.1',
     packages=['casim',
              ],
     license='MIT',
