### 2020-09-17 #2436 @CFGrote
I drafted the zenodo deposit, the reserved DOI is 10.5281/zenodo.4034071. I'm waiting my co-authors to give their "Ok" before publishing, will let you know.
BTW I bumped the version of cancer_sim to 2.0.1.
