### 2020-09-17 #2436 @pdebuyl
@CFGrote you can consider the latest pdf as a proof. Let me know when I can validate the publication. At that point, an editor in chief of JOSS will take over the publication.
