### 2020-09-17 mpievolbio-scicomp/cancer_sim#27 @CFGrote
Proof corrections
Addresses https://github.com/openjournals/joss-reviews/issues/2436#issuecomment-694118587

## TODOs:
- [x] Jake middle name inital
- [x] Check english spelling (UK)
- [x] Read and correct typos and misspellings
- [x] Check references
- [x] Check links
