### 2020-09-18 #2436 @CFGrote
> @CFGrote should I take this as confirmation to proceed?

please wait, i'm checking the binder link, last item on my to do list for the proofs.
