### 2020-09-18 mpievolbio-scicomp/cancer_sim#27 @CFGrote
Done (PR #28)
