### 2020-09-18 #2436 @CFGrote
@blakeaw , there's still one checkbox unchecked here (Review Checklist for blakeaw -> Functionality -> Functionality): https://github.com/openjournals/joss-reviews/issues/2436#issue-650881744 . Could you check, please?
