### 2020-09-18 #2436 @blakeaw
@CFGrote Ok, it should be checked. Let me know if it is registering as checked now.
(I unchecked that particular box and then re-checked it, as it was already registering as being checked on my end.)
