### 2020-09-18 #2436 @kyleniemeyer
Hi @CFGrote, I'm the EIC on duty this week, and doing some final checks before publishing.

In the paper, I noticed a few issues:
- In your title, I think "python3" should be changed to "Python 3"
- You do not have an explicit Statement of Need, one of our [required elements](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain), that illustrates the research purpose of the software (you can probably just reorganize your existing text slightly)
- please put commas after "i.e" or "e.g."
- in the third paragraph, I think you can remove the "(see e.g. Ref", and then the final outer ")", and just have the reference directly. With the nested parentheses it looks a little awkward right now.
