### 2020-09-22 #2436 @pdebuyl
@kyleniemeyer the paper was submitted before we required the "statement of need", which is why I didn't check for it (or whedon indeed).

Should the software archive be updated if only the paper is modified or can the author keep the current version 2.0.1 and the corresponding zenodo archive?
