### 2020-09-22 #2436 @kyleniemeyer
@pdebuyl no worries!

If only the software is modified, the archive does not need to be updated, since we (JOSS) archive the software.
