### 2020-09-23 2eb8a65
Addressing EIC @comments in https://github.com/openjournals/joss-reviews/issues/2436#issuecomment-695065949
