### 2020-09-23 c494f1f
Adding statement of need to readme (Addressing EIC @comments in https://github.com/openjournals/joss-reviews/issues/2436#issuecomment-695065949)
