### 2020-09-23 #2436 @kyleniemeyer
I see the paper at https://joss.theoj.org/papers/10.21105/joss.02436, but the DOI does not seem to be resolving. I'll hold off closing this until it does.
