### 2020-09-24 #2436 @kyleniemeyer
But, in the meantime, congrats @CFGrote on your article's publication in JOSS!

Many thank to @jenzopr and @blakeaw for reviewing this, and @pdebuyl for editing.
